extract grc finding builders
